data structures,greedy,*1100

+++ Codeforces Round 805 (Div. 3)/C. Train and Queries.cpp
@@ -63,7 +63,7 @@ void solve(void){
     
     ll n; cin >> n;
 
-    cout << sqrt(n);
+    cout << (ll)sqrt(n);
 
     nl;
 }
